Namespace some DWP-related code
Move the 3 service objects we currently have that deals with the DWP benefit checker under the `DWP` namespace.

This looks like a lot but is just renaming a few modules.

Code becomes much more manageable and easy to find if everything is kept consistent in namespaces when multiple classes relate to the same functionality/domain.
