+++ app/controllers/steps/dwp/confirm_result_controller.rb
@@ -1,6 +1,6 @@
 module Steps
-  module Dwp
-    class ConfirmResultController < Steps::DwpStepController
+  module DWP
+    class ConfirmResultController < Steps::DWPStepController
       def edit
         @form_object = ConfirmResultForm.new(
           crime_application: current_crime_application
