--- app/services/benefit_check_service.rb
@@ -1,8 +1,0 @@
-class BenefitCheckService
-  def self.call(applicant)
-    # For now we will just use the mock benefit check service
-    # and implement the full benefit check service functionality later
-
-    MockBenefitCheckService.passporting_benefit?(applicant)
-  end
-end
