+++ spec/controllers/steps/dwp/benefit_check_result_exit_controller_spec.rb
@@ -1,5 +1,5 @@
 require 'rails_helper'
 
-RSpec.describe Steps::Dwp::BenefitCheckResultExitController, type: :controller do
+RSpec.describe Steps::DWP::BenefitCheckResultExitController, type: :controller do
   it_behaves_like 'a show step controller'
 end
