+++ spec/controllers/steps/dwp/confirm_details_controller_spec.rb
@@ -1,5 +1,5 @@
 require 'rails_helper'
 
-RSpec.describe Steps::Dwp::ConfirmDetailsController, type: :controller do
-  it_behaves_like 'a generic step controller', Steps::Dwp::ConfirmDetailsForm, Decisions::DwpDecisionTree
+RSpec.describe Steps::DWP::ConfirmDetailsController, type: :controller do
+  it_behaves_like 'a generic step controller', Steps::DWP::ConfirmDetailsForm, Decisions::DWPDecisionTree
 end
