+++ spec/controllers/steps/dwp/confirm_result_controller_spec.rb
@@ -1,5 +1,5 @@
 require 'rails_helper'
 
-RSpec.describe Steps::Dwp::ConfirmResultController, type: :controller do
-  it_behaves_like 'a generic step controller', Steps::Dwp::ConfirmResultForm, Decisions::DwpDecisionTree
+RSpec.describe Steps::DWP::ConfirmResultController, type: :controller do
+  it_behaves_like 'a generic step controller', Steps::DWP::ConfirmResultForm, Decisions::DWPDecisionTree
 end
