+++ spec/services/dwp/mock_benefit_check_service_spec.rb
@@ -1,8 +1,6 @@
 require 'rails_helper'
 
-RSpec.describe MockBenefitCheckService do
-  subject { described_class }
-
+RSpec.describe DWP::MockBenefitCheckService do
   let(:last_name) { 'Smith' }
   let(:date_of_birth) { '1999/01/11'.to_date }
   let(:nino) { 'NC123459A' }
@@ -17,23 +15,25 @@ RSpec.describe MockBenefitCheckService do
   end
 
   describe '.call' do
+    subject { described_class.call(applicant) }
+
     it 'returns confirmation_ref' do
-      expect(subject.call(applicant)[:confirmation_ref]).to eq("mocked:#{described_class}")
+      expect(subject[:confirmation_ref]).to eq("mocked:#{described_class}")
     end
 
     it "returns 'Yes' as in known data" do
-      expect(subject.call(applicant)[:benefit_checker_status]).to eq('Yes')
+      expect(subject[:benefit_checker_status]).to eq('Yes')
     end
 
     context 'with incorrect date' do
       let(:date_of_birth) { '2012/01/10'.to_date }
 
       it 'returns no' do
-        expect(subject.call(applicant)[:benefit_checker_status]).to eq('No')
+        expect(subject[:benefit_checker_status]).to eq('No')
       end
 
       it 'returns confirmation_ref' do
-        expect(subject.call(applicant)[:confirmation_ref]).to eq("mocked:#{described_class}")
+        expect(subject[:confirmation_ref]).to eq("mocked:#{described_class}")
       end
     end
 
@@ -41,12 +41,14 @@ RSpec.describe MockBenefitCheckService do
       let(:last_name) { 'Unknown' }
 
       it 'returns no' do
-        expect(subject.call(applicant)[:benefit_checker_status]).to eq('No')
+        expect(subject[:benefit_checker_status]).to eq('No')
       end
     end
   end
 
   describe '.passporting_benefit?' do
+    subject { described_class }
+
     it 'returns true' do
       expect(subject.passporting_benefit?(applicant)).to be(true)
     end
